Backend/requirements.txt
fastapi==0.110.0          # ✅ Compatible with Pydantic v1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
motor==3.3.2
pymongo==4.6.0
python-decouple==3.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
razorpay==1.4.2           # ✅ last available release
pydantic[email]==1.10.13  # ✅ Pydantic v1.x (with email extras)
